requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docx2txt==0.8
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.2
joblib==0.16.0
lxml==4.5.2
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
